changes on orders

+++ src/main/java/com/ari/webapp/dto/ProductDto.java
@@ -6,19 +6,21 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 public class ProductDto {
     private Long id;
     private String name;
-    private Double price;
+    private BigDecimal price;
     private String image;
     private String description;
     private Integer stock;
     private Category category;
 
-    public ProductDto(String name, Double price, String image, String description, Integer stock, Category category) {
+    public ProductDto(String name, BigDecimal price, String image, String description, Integer stock, Category category) {
         this.name = name;
         this.price = price;
         this.image = image;
